Removed ConnectionExpired exception.
Removed strange state with defined with the ConnectionExpired exception.
